fix(scout-for-lol): surface the failure that wedges the Temporal supervisor
Why

Scout beta's Temporal supervisor spent 15 hours logging one identical warning
every ~6 seconds — "Temporal component is degraded; reconnecting", attempt 334,
"Cannot close connection while Workers hold a reference to it" — and that
message was never the failure. It came from the cleanup handler:

    } catch (error: unknown) {
      if (clientConnection !== undefined) await clientConnection.close();
      await nativeConnection.close();   // throws: workers still hold refs
      throw error;                      // unreachable
    }

`Worker.create` registers a reference holder on the native connection as soon
as it returns and only releases it once that worker's `run()` promise settles.
So any `Worker.create` failing part-way through the sequence left holders in
place, `close()` threw IllegalStateError, and the handler discarded the real
error along with it. Every attempt also leaked a connection, its native
workers, and a freshly built workflow bundle, which is the likely reason later
attempts failed where the first one succeeded.

What

- release a partially built runtime correctly: track workers as they are
  created, run and drain them, then close. The drain is bounded — a worker
  that will not settle would otherwise wedge the reconnect loop forever, which
  is worse than the leak it replaces. Cleanup failures are logged and
  swallowed so the caller's original error survives.
- push each worker individually. `push(await a(), await b())` evaluates both
  before pushing either, so a failure in the second orphans the first outside
  the array the cleanup path walks.
- make the legacy-namespace drain best-effort. Those pollers only let
  pre-cutover histories finish; one failing must not take the five live
  workers down with it.
- back off exponentially to a 60s cap instead of retrying flat every 5s, and
  log the stack and cause plus report to Sentry. The message alone was all
  that got logged, 334 times.
- split the runtime lifecycle into connected-runtime.ts. supervisor.ts had
  grown past the 500-line cap, and the class is orchestration while
  construction and teardown are their own concern.

Verification

- bunx turbo run typecheck lint --filter=@scout-for-lol/backend
  (architecture: 878 modules clean)
- vitest src/temporal/ src/startup.test.ts — 37 tests passed, including 7 new
  ones whose fake models the SDK's reference-holder invariant, so a release
  path that closed too early would fail them
- not yet observed live: needs an image build to reach the beta cluster
